auto archive and more
Added in Auto Archive of Imported Files.
Starting work on UseDefaultImportValues flag.
Will use Default values for all Auto Imported files. Don't want to
bombard the User with a crap tone of pop ups. Might make something that
will have a CSV with the files and the titles/notes that you want for
that file to be used for the auto import.
Thinking of how to do file logging.
Streamlined code for the folder import function. it is now a foreach
loop instead of having 4 sets of the same code. now i can just add a new
extension to the array instead of adding a whole new block of code.