Lots of changes
Added in settings for the Background threads.
Enable/Disable of the Daemon Tab and of Import Updates
Seconds interval for how often the thread runs.

Message box for single file import if it is already imported. Probably
will change this to just throw it in the listview.

Error in GetImportRows() fixed where it was not resetting the object to
be returned.