Some bug fixes
Working on trying to get performance to be better with very very large
folder sizes. Lots of files (about 1000+ files) starts to slow down the
UI updates. Need to figure out how to make it faster. the only way I
have right now is sleeps on the background thread and make the timer
longer. its at 8 minutes now and the bgthreads are at 1second sleep.

Added try/catches around API calls and on a few other things to prevent
crashing of the program. it is then logged to the Trace log (if
enabled).